section #29

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,8 @@ use App\Models\BlogPost;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use App\Models\Comment;
+use Illuminate\Support\Facades\Gate;
+use App\Models\User;
 
 class PostsController extends Controller
 {
@@ -36,7 +38,12 @@ class PostsController extends Controller
         // comments_count
         return view(
             'posts.index',
-            ['posts' => BlogPost::withCount('comments')->get()]
+            [
+                'posts' => BlogPost::latest()->withCount('comments')->get(),
+                'mostCommented' => BlogPost::mostCommented()->take(5)->get(),
+                'mostActive' => User::withMostBlogPosts()->take(5)->get(),
+                'mostActiveLastMonth' => User::withMostBlogPostsLastMonth()->take(5)->get(),
+            ]
         );
     }
 
@@ -47,6 +54,7 @@ class PostsController extends Controller
      */
     public function create()
     {
+        //$this->authorize('posts.create');
         return view('posts.create');
     }
 
@@ -59,6 +67,7 @@ class PostsController extends Controller
     public function store(StorePost $request)
     {
         $validated = $request->validated();
+        $validated['user_id'] = $request->user()->id;
         $post = BlogPost::create($validated);
 
         // $post = new BlogPost();
@@ -86,9 +95,14 @@ class PostsController extends Controller
     public function show($id)
     {
         // abord_if(!isset($this->posts[$id]), 404);
+        // return view('posts.show', [
+        //     'post' => BlogPost::with(['comments' => function ($query) {
+        //         return $query->latest();
+        //     }])->findOrFail($id),
+        // ]);
         return view('posts.show', [
-            'post' => BlogPost::with('comments')->findOrFail($id)]
-        );
+            'post' => BlogPost::with('comments')->findOrFail($id),
+        ]);        
     }
 
     /**
@@ -99,7 +113,15 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
-        return view('posts.edit', ['post' => BlogPost::findOrFail($id)]);
+        $post = BlogPost::findOrFail($id);
+
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't edit this post!");
+        // }  
+        
+        $this->authorize($post);
+        
+        return view('posts.edit', ['post' => $post]);
     }
 
     /**
@@ -112,6 +134,13 @@ class PostsController extends Controller
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't edit this post!");
+        // }
+
+        $this->authorize($post);
+
         $validated = $request->validated();
         $post->fill($validated);
         $post->save();
@@ -130,6 +159,13 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+
+        // if (Gate::denies('delete-post', $post)) {
+        //     abort(403, "You can't delete this post!");
+        // }
+
+        $this->authorize($post);
+
         $post->delete();
 
         session()->flash('status', 'Blog post was deleted!');
